Updated Red Example Score materials/ initializer.
Only temporary.

This commit makes the materials/__init__.py inert again, just to allow all IDE
tests to pass.

Explanation: running the IDE battery (unintentionally) changes the Red Example
Score materials/__init__.py from robust back to inert. This is a mistake. The
most likely cause is that the filesystem preservation context manager isn't be
used somewhere that it should be. This commit just accepts the old, inert
initializer, which will be fine for quite a while.