Improved segment metadata management.
The change is in boilerplate/__illustrate_segment__.py:

OLD:

    try:
        lilypond_file, sticky_settings = segment_maker(metadata=metadata)
    except:
        traceback.print_exc()
        sys.exit(1)

    try:
        metadata.update(sticky_settings)
        # write metadata to disk

NEW:

    try:
        lilypond_file, metadata = segment_maker(metadata=metadata)
    except:
        traceback.print_exc()
        sys.exit(1)

    try:
        # write metadata to disk

The old way of doing things relied on the segment-maker to return only a sticky
settings dictionary.

The new way of doing things elies on the segment-maker to return the enter
segment metadata dictionary.

The motivation here is to give the segment-maker absolute control over changing
(and even removing) segment metadata at call time.

Custom segment-makers should all now return a pair of a LilyPond file together
with a complete segment metadata dictionary, as shown above.