Added __aliases__.py properties to Configuration. NEW.
NEW. You can now add the file ...

    ~/.abjad/ide/__aliases__.py

... to your Abjad directory.

The file should contain a single ordered dictionary like this:

    # -*- encoding: utf-8 -*-
    import collections

    aliases = collections.OrderedDict([
        ('map', 'etc/map.pdf'),
        ])

The IDE will soon know how to interpret custom commands like (map) as shortcuts
to open the score/etc/map.pdf file *in any score*.

So the point will be to allow generalized navigation aliases that will work in
all scores belonging to a composer (as a type of site configuration).

The functionality is inert as of this commit. The functionality will be
activated in a future commit.