Taught IDE to ignore cache.
Time to run 840 tests with cache: 260 seconds.
Time to run 840 tests without cache: 297 seconds.

Explanation. IDE user experience was slow when users would enter
(ss) to return to all scores. This was true more than a year ago. But modern
versions of the IDE are very responsive. This means that the cache has not been
necessary to ensure IDE user experience for some time. The remaining reason for
keeping the cache was to make the tests run somewhat faster. The timings above
show that the cache does still confer a performance increase in the tests, but
not much of one. So this commit stops the ScorePackageWrangler from reading the
cache. Future commit will remove the cache and cache-management commands
completely.