Refactored AbjadIDEConfiguration access. PERFORMANCE.
AbjadIDEConfiguration is now instantiated as a module-global object in the
objects that require AbjadIDEConfiguration access.

OLD. AbjadIDEConfiguration was instantiated as an *instance* variable on all
controllers. This means that file system reads to ~/.abjad/ide/ide.cfg happened
on the initialization on every controller. Tests ran in approximately 90
seconds.

NEW. AbjadIDEConfiguration is now instantiated as a module-global object in
only those modules that require AbjadIDEConfiguration access. Tests run in
approximately 50 seconds.

PERFORMANCE. Test run time decreases from 90 seconds to 50 seconds with this
commit.