Added new navigation behavior. NEW.
The behavior is currently implemented only for file aliases.

DESCRIPTION. Two different ways of editing files come up over and over again
when working inside an Abjad score package.

    1. You're working somewhere in a score package and you want temporarily to
    open a file, make changes and then resume working in the directory where
    you were working before opening the file. This is the behavior implemented
    by 'stylesheet - edit' (sty) and 'log - edit' (log). The thing to note here
    is that first 'working context' (ie, directory within the score package)
    should not be disturbed by the act of temporarily viewing a reference file.

    2. You're working somewhere in a score package and you want to move to
    to another directory in the score pakcage, open a file, make changes and
    then begin working in this new directory where the file is housed. This is
    what we normally do when, for example, we're in the makers/ directory and
    we realize that we wnat to "pop over" to segment #3, edit the definition
    file and then illustrate the definition. Composers must currently enact the
    steps in this pattern by hand with something like (gg segment 03 df pdfm).
    The thing to note here is that this is basically a "definition edit"
    pattern: we execute this pattern when we realize -- in the middle of
    working -- that we want to edit (and subsequently check or illustrate) a
    "remote" definition file for a material or a segment.

This commit takes a first step towards implementing the navigation pattern
described above in #2.

As of this commit, aliases that resolve to a file in a material directory or in
a segment directory will first open the file for editing (as expected) and then
manage the material directory or segment directory housing the file.

Note that aliases that resolve to a file housed in any other directory (build,
distribution, etc, makers, stylesheets) will open the file for editing but
return the composer to the previous working directory after file close.

Further commits will extend this "edit-and-change" pattern to other
interactions in the system.