Taught IDE about preexisting empty score directories. NEW.
Calling (new) to create a new score now causes the IDE to the composer's score
directory for any empty directories. The IDE will then ask the composer if the
empty directory should be treated as a score directory and populated with
necessary contents:

    Abjad IDE - all score directories

    1: Blue Example Score (2013)
    2: Red Example Score (2013)
    3: foo

        copy (cp)
        new (new)
        remove (rm)
        rename (ren)

    > new
    Found empty score directory foo/.
    Populate foo/ with score contents? y

Why is this useful? The point of the extension is to allow composers to create
and clone an (empty) repository on GitHub and subsequently to populate the
local copy of the repository using the IDE. This streamlines score package
creation considerably.