Refactored segment (pdfm).
OLD. THe IDE removed the __illustrate__.py file that is created when making a
segment illustration.pdf with (pdfm).

NEW. The IDE now preserves the __illustrate__.py file that is created when
making a segment illustration.pdf with (pdfm).

NOTE. This means that the next time you remake PDFs for an entire score with
(pdfm*) or equivalent, a series of __illustrate__.py files will appear in your
segment directories.

The thinking here is that material (pdfm) already relies on the presence of a
(user-created) __ilustrate__.py file. So this commit harmonizes segment (pdfm)
and material (pdfm).

The decision adds another artifact to segment directories. But the presence of
the __illustrate__.py file makes it possible for users to customize segment
illustration (although this should be unnecessary). It also makes it much
easier to debug segments illustration.

Basically the intuition was that the IDE was performing a nontrivial type of
work in more-or-less undocumented way.

The work is now clearly documented as it happens.